시리얼라이저 구현 및 그에 따른 리펙터링
