+++ .gitignore
@@ -1,3 +1,4 @@
-/out/
-/build/
-/.gradle/
+/out
+/build
+/.gradle
+/src/main/java/Test.java
