+++ src/main/java/persistence/enums/RegistStatus.java
@@ -1,13 +1,15 @@
 package persistence.enums;
 
 import lombok.Getter;
+
+import java.io.Serializable;
 import java.util.Collections;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Getter
-public enum RegistStatus implements Enum {
+public enum RegistStatus implements Enum, Serializable {
     HOLD("보류"),
     ACCEPT("수락"),
     REJECT("거절");
